add report, create request script
