+++ explorer/imports/ui/components/msg/MsgDetail.jsx
@@ -5,25 +5,46 @@ import MsgTable from './MsgTable'
 
 const mockData = [
   {
+    action: 'CREATE_SCRIPT',
     scriptHash: '0x9q391923912392139231923912323',
     creator: '0x01230013012023021301230',
     name: 'ETH/USD Median Price',
     sources: ['CoinMarketcap', 'Crypto Compare', 'Binance'],
     tag: ['Crypto Price'],
+    code: 'You will see the code right here in the future.',
   },
   {
-    scriptHash: '0x834473473473467347347347347',
-    creator: '0x1348421874384384383484383483',
-    name: 'BTC/USD Median Price',
-    sources: ['CoinMarketcap', 'Crypto Compare'],
-    tag: ['Crypto Price'],
+    action: 'REPORT',
+    reqID: '32',
+    status: 'Succes',
+    from: '0x1348421874384384383484383483',
+    data: ['0x0000008332', '0x0000008332'],
   },
   {
-    scriptHash: '0x834473473473467347347347347',
-    creator: '0x1348421874384384383484383483',
-    name: 'BTC/USD Median Price',
-    sources: ['CoinMarketcap', 'Crypto Compare'],
-    tag: ['Crypto Price'],
+    action: 'REQUEST',
+    reqID: '32',
+    status: 'pending',
+    ts: '1,329',
+    from: '0x1348421874384384383484383483',
+    script: '0x1348421874384384383484383483',
+    result: '0x0000008332',
+    proof: 'asdkaksdkasdkaskdk',
+    reports: [
+      {
+        txHash: '0x9123912923192139213912391239329',
+        block: '1,324',
+        ts: '22',
+        from: '0x85334473743743743734734734734',
+        value: ['0x0000008332', '0x0000008332', '0x0000008332'],
+      },
+      {
+        txHash: '0x9123912923192139213912391239329',
+        block: '1,324',
+        ts: '22',
+        from: '0x85334473743743743734734734734',
+        value: ['0x0000008332', '0x0000008332', '0x0000008332'],
+      },
+    ],
   },
 ]
